Issue #412: Move CursorController to its own file in browser.cursor pkg.
